Finish client functionalities, fix socket bug
The assignment should now be finished. A few final changes were made,
most notably fixing a socket bug that gave a "Bad file descriptior" IO
exception when transferring large files.

The bug was that on each packet sent, and on every packet resend, a new
UDP socket was being created. This error was fixed by moving the socket
initialization to the constructor.

A large image file was also added for testing.